requirements.txt
beautifulsoup4
fusepy
requests
